add posts rating
